// IraqiTradeDashboard/src/types/api.ts
export interface CompanyDto {
  id: number;
  code: string;
  nameAr: string;
  phone: string;
  email?: string;
  city?: string;
  address?: string;
  contactPerson?: string;
  isActive: boolean;
  subscriptionExpiry?: string;
  createdAt: string;
}